feat: move repository out as separate module
