+++ backend/solontax-g1-management-dao/pom.xml
@@ -4,9 +4,9 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0
                              http://maven.apache.org/xsd/maven-4.0.0.xsd">
 
-    <artifactId>solontax-g1-management-core</artifactId>
+    <artifactId>solontax-g1-management-dao</artifactId>
     <modelVersion>4.0.0</modelVersion>
-    <name>solontax-g1-management-core</name>
+    <name>solontax-g1-management-dao</name>
     <packaging>jar</packaging>
 
     <parent>
@@ -54,30 +54,26 @@
         </dependency>
 
         <dependency>
-            <artifactId>spring-boot-starter-validation</artifactId>
-            <groupId>org.springframework.boot</groupId>
-        </dependency>
-
-        <dependency>
-            <artifactId>spring-boot-starter-data-jpa</artifactId>
-            <groupId>org.springframework.boot</groupId>
-        </dependency>
-
-        <dependency>
             <artifactId>spring-boot-starter-actuator</artifactId>
             <groupId>org.springframework.boot</groupId>
         </dependency>
 
         <dependency>
-            <artifactId>rsql-parser</artifactId>
-            <groupId>cz.jirutka.rsql</groupId>
-            <version>2.1.0</version>
+            <groupId>org.projectlombok</groupId>
+            <artifactId>lombok</artifactId>
+            <scope>annotationProcessor</scope>
         </dependency>
 
         <dependency>
-            <artifactId>lombok</artifactId>
-            <groupId>org.projectlombok</groupId>
-            <optional>true</optional>
+            <groupId>solontax.g1</groupId>
+            <artifactId>solontax-g1-management-core</artifactId>
+            <version>0.0.1-SNAPSHOT</version>
+        </dependency>
+
+        <dependency>
+            <groupId>org.postgresql</groupId>
+            <artifactId>postgresql</artifactId>
+            <scope>runtime</scope>
         </dependency>
     </dependencies>
 </project>
