+++ backend/solontax-g1-management-migration/pom.xml
@@ -68,11 +68,23 @@
             <artifactId>lombok</artifactId>
             <scope>annotationProcessor</scope>
         </dependency>
+
+        <dependency>
+            <groupId>com.fasterxml.jackson.datatype</groupId>
+            <artifactId>jackson-datatype-jsr310</artifactId>
+            <version>2.17.2</version> <!-- or match your jackson-databind version -->
+        </dependency>
+
         <dependency>
             <groupId>solontax.g1</groupId>
             <artifactId>solontax-g1-management-core</artifactId>
             <version>0.0.1-SNAPSHOT</version>
-            <scope>compile</scope>
+        </dependency>
+
+        <dependency>
+            <groupId>solontax.g1</groupId>
+            <artifactId>solontax-g1-management-dao</artifactId>
+            <version>0.0.1-SNAPSHOT</version>
         </dependency>
     </dependencies>
 </project>
