--- backend/solontax-g1-management-migration/src/test/java/solontax/g1/management/migration/SolontaxG1ManagementMigrationApplicationTests.java
@@ -1,13 +1,0 @@
-package solontax.g1.management.migration;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SolontaxG1ManagementMigrationApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
